Feature: Design system, FooterNav (#653)
* clean up tailwind config

* fix border radius

* refactor Logo

* init tw footerNav

* convert sass to TW

* replace old FooterNavs

* delete old FooterNav
