--- frontend/src/components/Navigation/_FooterNav.scss
@@ -1,94 +1,0 @@
-@use "../Basics" as *;
-
-.footer-nav {
-  background-color: $color-blue-darker;
-  height: 128px;
-  align-items: center;
-
-  @include breakpoint-media-max("lgtablet") {
-    height: 226px;
-    flex-direction: column;
-    justify-content: space-evenly;
-  }
-}
-
-.logo-desktop-footer {
-  @include breakpoint-media-max("mobile") {
-    @include hidden;
-  }
-}
-
-.logo-mobile-footer {
-  @include hidden;
-  @include breakpoint-media-max("mobile") {
-    display: block;
-  }
-}
-
-.footer-icons-on-dark {
-  display: block;
-  margin-left: 176px;
-
-  @include breakpoint-media-max("lgtablet") {
-    margin-left: 0;
-    margin-bottom: -17px;
-  }
-}
-
-.footer-menu {
-  @include breakpoint-media-max("lgtablet") {
-    width: 220px;
-    order: 3;
-  }
-}
-
-.footer-donate-button {
-  margin-left: auto;
-  margin-right: 188px;
-
-  @include breakpoint-media-max("lgtablet") {
-    margin-left: 0;
-    margin-right: 0;
-  }
-}
-
-.footer-menu-vertical-line {
-  border: 1px solid $color-white;
-  width: 40px;
-  transform: rotate(-90deg);
-  margin: auto;
-
-  &:first-child {
-    display: block;
-  }
-
-  &:not(:first-child) {
-    visibility: hidden;
-  }
-
-  @include breakpoint-media-max("lgtablet") {
-    width: 20px;
-
-    &:first-child {
-      @include hidden;
-    }
-
-    &:not(:first-child) {
-      visibility: visible;
-    }
-  }
-}
-
-.footer-links {
-  color: $color-white;
-  text-decoration: none;
-  @include title(6);
-
-  @include breakpoint-media-max("mobile") {
-    @include title(7);
-  }
-
-  &:hover {
-    text-decoration: underline;
-  }
-}
