+++ frontend/src/components/Navigation/_index.scss
@@ -1,2 +1,1 @@
-@forward "FooterNav";
 @forward "ProgressBar";
