+++ frontend/src/layouts/DefaultNavLayout.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import { FooterNav } from "components/components";
-import { HeaderNav } from "tw-components";
+import { HeaderNav, FooterNav } from "tw-components";
 
 const DefaultNavLayout = () => {
   return (
