+++ frontend/src/tw-components/index.tsx
@@ -1,4 +1,5 @@
 export { default as AuthNav } from "./AuthNav";
-export { default as HeaderNav } from "./HeaderNav";
 export { default as CookieBanner } from "./CookieBanner";
+export { default as FooterNav } from "./FooterNav";
+export { default as HeaderNav } from "./HeaderNav";
 export { default as TextField } from "./TextField";
