+++ frontend/tailwind.config.js
@@ -1,6 +1,11 @@
 /** @type {import('tailwindcss').Config} */
+
+/*
+ * Please refer to CTJ Figma Design System for more details about custom theme values
+ **/
+
 module.exports = {
-  darkMode: "class", // Enables dark mode with a class
+  darkMode: "class",
   content: [
     "./src/pages/Demo/DemoTailwind.tsx",
     "./src/pages/Authentication/*.tsx",
@@ -12,15 +17,10 @@ module.exports = {
   theme: {
     screens: {
       xs: "480px",
-      // => @media (min-width: 480px) { ... }
       sm: "577px",
-      // => @media (min-width: 577px) { ... }
       md: "769px",
-      // => @media (min-width: 769px) { ... }
       lg: "1025px",
-      // => @media (min-width: 1025px) { ... }
       xl: "1201px",
-      // => @media (min-width: 1201px) { ... }
     },
     colors: {
       // Primary Colors
@@ -37,13 +37,13 @@ module.exports = {
       "tan-light": "#ffefdb",
       "tan-bg": "#fbe8ce",
       green: "#13831e",
-      red: "#CC0023",
+      red: "#c93329",
       // Neutral Colors
-      white: "#fff",
+      white: "#ffffff",
       "grey-light": "#f2f2f2",
       grey: "#c1c1c1",
       "grey-dark": "#585858",
-      charcoal: "#333",
+      charcoal: "#333333",
     },
     fontFamily: {
       sans: ["Roboto", "Tahoma", "Verdana", "sans-serif"],
@@ -57,17 +57,30 @@ module.exports = {
       black: "900",
     },
     extend: {
-      lineHeight: {
-        "extra-tight": "1.125rem",
-      },
       borderRadius: {
         DEFAULT: "20px",
         large: "60px",
         "x-large": "100px",
       },
+      lineHeight: {
+        "extra-tight": "1.125rem",
+      },
       rotate: {
         290: "290deg",
         345: "345deg",
+      },
+      spacing: {
+        p0: "0px",
+        p1: "8px",
+        p2: "16px",
+        p3: "24px",
+        p4: "32px",
+        p5: "40px",
+        p6: "48px",
+        p7: "56px",
+        p8: "64px",
+        p9: "72px",
+        p10: "80px",
       },
     },
   },
